nanobot-video-generator/vision/pyproject.toml
[project]
name = "mcp-server-vision"
version = "0.0.1"
description = "A Model Context Protocol server providing tools to analyze images"
readme = "README.md"
requires-python = ">=3.13"
authors = [{ name = "drpebcak" }]
keywords = ["http", "mcp", "llm", "automation"]
license = { text = "MIT" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "fastmcp>=2.6.1",
    "openai>=1.84.0",
    "pillow>=11.2.1",
    "python-dotenv>=1.1.0",
    "httpx>=0.28.1",
]

[project.scripts]
mcp-server-vision = "mcp_server_vision:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = ["pyright>=1.1.389", "ruff>=0.11.12"]

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_server_vision"]